+++ src/main/java/org/jabref/gui/util/DefaultTaskExecutor.java
@@ -2,6 +2,8 @@ package org.jabref.gui.util;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.FutureTask;
 import java.util.function.Consumer;
 
@@ -19,9 +21,31 @@ public class DefaultTaskExecutor implements TaskExecutor {
 
     private static final Log LOGGER = LogFactory.getLog(DefaultTaskExecutor.class);
 
+    private ExecutorService executor = Executors.newFixedThreadPool(5);
+
+    public static <V> V runInJavaFXThread(Callable<V> callable) {
+        FutureTask<V> task = new FutureTask<>(callable);
+        Platform.runLater(task);
+        try {
+            return task.get();
+        } catch (InterruptedException | ExecutionException e) {
+            LOGGER.error(e);
+            return null;
+        }
+    }
+
+    public static void runInJavaFXThread(Runnable runnable) {
+        Platform.runLater(runnable);
+    }
+
     @Override
     public <V> void execute(BackgroundTask<V> task) {
-        new Thread(getJavaFXTask(task)).start();
+        executor.submit(getJavaFXTask(task));
+    }
+
+    @Override
+    public void shutdown() {
+        executor.shutdownNow();
     }
 
     private <V> Task<V> getJavaFXTask(BackgroundTask<V> task) {
@@ -54,16 +78,4 @@ public class DefaultTaskExecutor implements TaskExecutor {
             return new Exception(throwable);
         }
     }
-
-    public static <V> V runInJavaFXThread(Callable<V> callable) {
-        FutureTask<V> task = new FutureTask<>(callable);
-        Platform.runLater(task);
-        try {
-            return task.get();
-        } catch (InterruptedException | ExecutionException e) {
-            LOGGER.error(e);
-            return null;
-        }
-    }
-
 }