+++ src/main/java/org/jabref/logic/importer/fetcher/DoiResolution.java
@@ -43,7 +43,7 @@ public class DoiResolution implements FulltextFetcher {
                 try {
                     Connection connection = Jsoup.connect(sciLink);
                     // pretend to be a browser (agent & referrer)
-                    connection.userAgent("Mozilla/5.0 (Windows; U; WindowsNT 5.1; en-US; rv1.8.1.6) Gecko/20070725 Firefox/2.0.0.6");
+                    connection.userAgent(URLDownload.USER_AGENT);
                     connection.referrer("http://www.google.com");
                     connection.followRedirects(true);
                     connection.ignoreHttpErrors(true);