+++ src/main/java/org/jabref/logic/msbib/BibTeXConverter.java
@@ -16,6 +16,9 @@ public class BibTeXConverter {
 
     private static final String MSBIB_PREFIX = "msbib-";
 
+    private BibTeXConverter() {
+    }
+
     /**
      * Converts an {@link MSBibEntry} to a {@link BibEntry} for import
      * @param entry The MsBibEntry to convert
@@ -100,11 +103,11 @@ public class BibTeXConverter {
         return result;
     }
 
-    private static void addAuthor(Map<String, String> map, String type, List<PersonName> authors) {
+    private static void addAuthor(Map<String, String> map, String type, List<MsBibAuthor> authors) {
         if (authors == null) {
             return;
         }
-        String allAuthors = authors.stream().map(PersonName::getFullname).collect(Collectors.joining(" and "));
+        String allAuthors = authors.stream().map(MsBibAuthor::getLastFirst).collect(Collectors.joining(" and "));
 
         map.put(type, allAuthors);
     }