+++ src/main/java/org/jabref/logic/pdf/PdfAnnotationImporter.java
@@ -1,11 +1,10 @@
 package org.jabref.logic.pdf;
 
 import java.awt.geom.Rectangle2D;
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
@@ -13,10 +12,8 @@ import java.util.Locale;
 import java.util.Objects;
 import java.util.Optional;
 
-import org.jabref.logic.util.io.FileUtil;
-import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.pdf.FileAnnotation;
-import org.jabref.preferences.JabRefPreferences;
+import org.jabref.model.pdf.FileAnnotationType;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -26,8 +23,6 @@ import org.apache.pdfbox.cos.COSName;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
 import org.apache.pdfbox.pdmodel.common.PDRectangle;
-import org.apache.pdfbox.pdmodel.fdf.FDFAnnotationHighlight;
-import org.apache.pdfbox.pdmodel.fdf.FDFAnnotationText;
 import org.apache.pdfbox.pdmodel.interactive.annotation.PDAnnotation;
 import org.apache.pdfbox.util.PDFTextStripperByArea;
 
@@ -43,24 +38,25 @@ public class PdfAnnotationImporter implements AnnotationImporter {
      * @return a list with the all the annotations found in the file of the path
      */
     @Override
-    public List<FileAnnotation> importAnnotations(final Path path, final BibDatabaseContext context) {
+    public List<FileAnnotation> importAnnotations(final Path path) {
 
-        Optional<Path> validatePath = validatePath(path, context);
-        if (!validatePath.isPresent()) {
+        if (!validatePath(path)) {
             // Path could not be validated, return default result
             return Collections.emptyList();
         }
 
-        Path validPath = validatePath.get();
-
         List<FileAnnotation> annotationsList = new LinkedList<>();
-        try (PDDocument document = PDDocument.load(validPath.toString())) {
+        try (PDDocument document = PDDocument.load(path.toString())) {
             List pdfPages = document.getDocumentCatalog().getAllPages();
             for (int pageIndex = 0; pageIndex < pdfPages.size(); pageIndex++) {
                 PDPage page = (PDPage) pdfPages.get(pageIndex);
                 for (PDAnnotation annotation : page.getAnnotations()) {
-                    if (annotation.getSubtype().equals(FDFAnnotationHighlight.SUBTYPE)) {
-                        annotationsList.add(createHighlightAnnotations(pageIndex, page, annotation));
+                    if (!isSupportedAnnotationType(annotation)) {
+                        continue;
+                    }
+                    if (FileAnnotationType.UNDERLINE.toString().equals(annotation.getSubtype()) ||
+                            FileAnnotationType.HIGHLIGHT.toString().equals(annotation.getSubtype())) {
+                        annotationsList.add(createMarkedAnnotations(pageIndex, page, annotation));
                     } else {
                         FileAnnotation fileAnnotation = new FileAnnotation(annotation, pageIndex + 1);
                         if (fileAnnotation.getContent() != null && !fileAnnotation.getContent().isEmpty()) {
@@ -70,31 +66,45 @@ public class PdfAnnotationImporter implements AnnotationImporter {
                 }
             }
         } catch (IOException e) {
-            LOGGER.error(String.format("Failed to read file '%s'.", validPath), e);
+            LOGGER.error(String.format("Failed to read file '%s'.", path), e);
         }
         return annotationsList;
     }
 
-    private FileAnnotation createHighlightAnnotations(int pageIndex, PDPage page, PDAnnotation annotation) {
-        FileAnnotation annotationBelongingToHighlighting = new FileAnnotation(
-                annotation.getDictionary().getString(COSName.T), FileAnnotation.extractModifiedTime(annotation.getModifiedDate()),
-                pageIndex + 1, annotation.getContents(), FDFAnnotationText.SUBTYPE, Optional.empty());
-
+    private boolean isSupportedAnnotationType(PDAnnotation annotation) {
         try {
-            annotation.setContents(extractHighlightedText(page, annotation));
-        } catch (IOException e) {
-            annotation.setContents("JabRef: Could not extract any highlighted text!");
+            if (!Arrays.asList(FileAnnotationType.values()).contains(FileAnnotationType.valueOf(annotation.getSubtype()))) {
+                return false;
+            }
+        } catch (IllegalArgumentException e) {
+            LOGGER.debug(String.format("Could not parse the FileAnnotation %s into any known FileAnnotationType. It was %s!", annotation, annotation.getSubtype()));
         }
-
-        //highlighted text that has a sticky note on it should be linked to the sticky note
-        return new FileAnnotation(annotation, pageIndex + 1, annotationBelongingToHighlighting);
+        return true;
     }
 
-    private String extractHighlightedText(PDPage page, PDAnnotation annotation) throws IOException {
-        //highlighted text has to be extracted by the rectangle calculated from the highlighting
+    private FileAnnotation createMarkedAnnotations(int pageIndex, PDPage page, PDAnnotation annotation) {
+        FileAnnotation annotationBelongingToMarking = new FileAnnotation(
+                annotation.getDictionary().getString(COSName.T), FileAnnotation.extractModifiedTime(annotation.getModifiedDate()),
+                pageIndex + 1, annotation.getContents(), FileAnnotationType.valueOf(annotation.getSubtype().toUpperCase(Locale.ROOT)), Optional.empty());
+
+        try {
+            if (FileAnnotationType.HIGHLIGHT.toString().equals(annotation.getSubtype()) || FileAnnotationType.UNDERLINE.toString().equals(annotation.getSubtype())) {
+                annotation.setContents(extractMarkedText(page, annotation));
+            }
+        } catch (IOException e) {
+            annotation.setContents("JabRef: Could not extract any marked text!");
+        }
+
+        //Marked text that has a sticky note on it should be linked to the sticky note
+        return new FileAnnotation(annotation, pageIndex + 1, annotationBelongingToMarking);
+    }
+
+
+    private String extractMarkedText(PDPage page, PDAnnotation annotation) throws IOException {
+        //highlighted or underlined text has to be extracted by the rectangle calculated from the marking
         PDFTextStripperByArea stripperByArea = new PDFTextStripperByArea();
         COSArray quadsArray = (COSArray) annotation.getDictionary().getDictionaryObject(COSName.getPDFName("QuadPoints"));
-        String highlightedText = "";
+        String markedText = "";
         for (int j = 1,
              k = 0;
              j <= (quadsArray.size() / 8);
@@ -118,44 +128,38 @@ public class PdfAnnotationImporter implements AnnotationImporter {
             uly = pageSize.getHeight() - uly;
 
             Rectangle2D.Float rectangle = new Rectangle2D.Float(ulx, uly, width, height);
-            stripperByArea.addRegion("highlightedRegion", rectangle);
+            stripperByArea.addRegion("markedRegion", rectangle);
             stripperByArea.extractRegions(page);
-            String highlightedTextInLine = stripperByArea.getTextForRegion("highlightedRegion");
+            String markedTextInLine = stripperByArea.getTextForRegion("markedRegion");
 
             if (j > 1) {
-                highlightedText = highlightedText.concat(highlightedTextInLine);
+                markedText = markedText.concat(markedTextInLine);
             } else {
-                highlightedText = highlightedTextInLine;
+                markedText = markedTextInLine;
             }
         }
 
-        return highlightedText.trim();
+        return markedText.trim();
     }
 
-    private Optional<Path> validatePath(Path path, BibDatabaseContext context) {
+    private boolean validatePath(Path path) {
         Objects.requireNonNull(path);
 
         if (!path.toString().toLowerCase(Locale.ROOT).endsWith(".pdf")) {
-            LOGGER.warn(String.format("File %s does not end with .pdf!", path.toString()));
-            return Optional.empty();
+            LOGGER.warn(String.format("File %s does not end with .pdf!", path));
+            return false;
         }
 
         if (!Files.exists(path)) {
-            Optional<File> importedFile = FileUtil.expandFilename(context, path.toString(),
-                    JabRefPreferences.getInstance().getFileDirectoryPreferences());
-            if (importedFile.isPresent()) {
-                // No need to check file existence again, because it is done in FileUtil.expandFilename()
-                return Optional.of(Paths.get(importedFile.get().getAbsolutePath()));
-            } else {
-                //return empty list to recognize invalid import
-                return Optional.empty();
-            }
+            LOGGER.warn(String.format("File %s does not exist!", path));
+            return false;
         }
 
         if (!Files.isRegularFile(path) || !Files.isReadable(path)) {
-            LOGGER.warn(String.format("File %s is not readable!", path.toString()));
-            return Optional.empty();
+            LOGGER.warn(String.format("File %s is not readable!", path));
+            return false;
         }
-        return Optional.of(path);
+
+        return true;
     }
 }