+++ src/main/java/org/jabref/logic/util/MetadataSerializationConfiguration.java
@@ -57,4 +57,7 @@ public class MetadataSerializationConfiguration {
      * Identifier for {@link AutomaticKeywordGroup}.
      */
     public static final String AUTOMATIC_KEYWORD_GROUP_ID = "AutomaticKeywordGroup:";
+
+    private MetadataSerializationConfiguration() {
+    }
 }